+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+2.1.5
++++++
+* Allow idempotent create for managed registry.
+* Add a no-format flag for displaying build logs
+
 2.1.4
 +++++
 * Provide a workaround for runtime operations without ARM requests.